[TweetPublisher] decouple PostsProvider
