+++ src/TweetPublisher/TweetPublisher.php
@@ -67,7 +67,8 @@ final class TweetPublisher
         }
 
         $tweet = $this->pickTweetCandidate($tweetsToPublish);
-        $this->twitterApiWrapper->publishTweet($tweet);
+
+//        $this->twitterApiWrapper->publishTweet($tweet);
 
         $this->symfonyStyle->success(sprintf(
             'Tweet "%s" was succesfully published.',
